requirements.txt
asyncio==3.4.3
telethon==1.29.3
aiofiles==23.2.1
aiohttp==3.8.5
numpy==1.25.2
pillow==10.0.0
psutil
pyrogram==2.0.106
py-tgcalls==0.9.7
python-dotenv==1.0.0
tgcrypto==1.2.5
speedtest-cli==2.1.3
git+https://github.com/yt-dlp/yt-dlp@master
git+https://github.com/alexmercerind/youtube-search-python@main
git+https://github.com/joetats/youtube_search@master
